add multipart rom array's

+++ avr/usbload/rle.h
@@ -23,6 +23,6 @@
 
 #include <avr/pgmspace.h>
 
-uint8_t rle_decode(PGM_VOID_P in_addr, uint32_t in_len, uint32_t out_addr);
+uint32_t rle_decode(PGM_VOID_P in_addr, uint32_t in_len, uint32_t out_addr);
 
 #endif
